Add @Description MetaBlock and other thing
The other thing is just add function, to work only if the user not typing. If typing, the script will be ignored. If not, the script will run and do it business😎
Oh yeah, and I also change the shortcut to " ` " only because I wanna test if it's work or not. If yes than Cool if not than shit :P