Little change to new beginning, Don't forgot me please!
A lot of thing just changed..
although, visually you only see the Name and text just changed a bit--but actually a lot of thing just changed(:
This meant to just add focus shortcut to the field, I just wanted this to be individually function. But I change my mind so.. It's combined now:)
I can't wait to see what the future looks like!