Updating the comments selector..
I just realized that a comment has been hidden and was replaced by a button to show them all, which has classes similar to those that I use to select all comments. Except that it has extra classes, so I just exclude the element with it.
I also included some log information in the title attribute, so that the anyone using it is informed that some comments are not being scanned. To make it scannable, you must always reveal the comment by yourself (simply press by using the button).